refactor: move slayer logic from plugin to notifier
