+++ src/main/java/dinkplugin/notifiers/SlayerNotifier.java
@@ -5,14 +5,15 @@ import dinkplugin.message.NotificationBody;
 import dinkplugin.message.NotificationType;
 import dinkplugin.Utils;
 import dinkplugin.notifiers.data.SlayerNotificationData;
-import lombok.Getter;
-import lombok.Setter;
 
 import javax.inject.Inject;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
-@Getter
-@Setter
 public class SlayerNotifier extends BaseNotifier {
+    public static final Pattern SLAYER_TASK_REGEX = Pattern.compile("You have completed your task! You killed (?<task>[\\d,]+ [^.]+)\\..*");
+    private static final Pattern SLAYER_COMPLETE_REGEX = Pattern.compile("You've completed (?:at least )?(?<taskCount>[\\d,]+) (?:Wilderness )?tasks?(?: and received (?<points>\\d+) points, giving you a total of [\\d,]+|\\.You'll be eligible to earn reward points if you complete tasks from a more advanced Slayer Master\\.| and reached the maximum amount of Slayer points \\((?<points2>[\\d,]+)\\))?");
+
     private String slayerTask = "";
     private String slayerPoints = "";
     private String slayerCompleted = "";
@@ -48,5 +49,40 @@ public class SlayerNotifier extends BaseNotifier {
         slayerTask = "";
         slayerPoints = "";
         slayerCompleted = "";
+    }
+
+    public void onChatMessage(String chatMessage) {
+
+        if (config.notifySlayer()
+            && (chatMessage.contains("Slayer master")
+            || chatMessage.contains("Slayer Master")
+            || chatMessage.contains("completed your task!")
+        )) {
+            Matcher taskMatcher = SLAYER_TASK_REGEX.matcher(chatMessage);
+            Matcher pointsMatcher = SLAYER_COMPLETE_REGEX.matcher(chatMessage);
+
+            if (taskMatcher.find()) {
+                this.slayerTask = taskMatcher.group("task");
+                this.handleNotify();
+            }
+
+            if (pointsMatcher.find()) {
+                String slayerPoints = pointsMatcher.group("points");
+                String slayerTasksCompleted = pointsMatcher.group("taskCount");
+
+                if (slayerPoints == null) {
+                    slayerPoints = pointsMatcher.group("points2");
+                }
+
+                // 3 different cases of seeing points, so in our worst case it's 0
+                if (slayerPoints == null) {
+                    slayerPoints = "0";
+                }
+                this.slayerPoints = slayerPoints;
+                this.slayerCompleted = slayerTasksCompleted;
+
+                this.handleNotify();
+            }
+        }
     }
 }
